+++ bigbio/biodatasets/drugprot/drugprot.py
@@ -22,7 +22,7 @@ https://biocreative.bioinformatics.udel.edu/tasks/biocreative-vii/track-1/
 """
 import collections
 from pathlib import Path
-from typing import Dict, Iterator, Tuple, Optional
+from typing import Dict, Iterator, Tuple
 
 import datasets
 
@@ -140,44 +140,32 @@ class DrugProtDataset(datasets.GeneratorBasedBuilder):
         return [
             datasets.SplitGenerator(
                 name=datasets.Split.TRAIN,
-                gen_kwargs={
-                    "abstracts_file": data_dir / "training" / "drugprot_training_abstracs.tsv",
-                    "entities_file": data_dir / "training" / "drugprot_training_entities.tsv",
-                    "relations_file": data_dir / "training" / "drugprot_training_relations.tsv",
-                },
+                gen_kwargs={"data_dir": data_dir, "split": "training"},
             ),
             datasets.SplitGenerator(
                 name=datasets.Split.VALIDATION,
-                gen_kwargs={
-                    "abstracts_file": data_dir / "development" / "drugprot_development_abstracs.tsv",
-                    "entities_file": data_dir / "development" / "drugprot_development_entities.tsv",
-                    "relations_file": data_dir / "development" / "drugprot_development_relations.tsv",
-                },
-            ),
-            datasets.SplitGenerator(
-                name=datasets.Split("test_background"),
-                gen_kwargs={
-                    "abstracts_file": data_dir / "test-background" / "test_background_abstracts.tsv",
-                    "entities_file": data_dir / "test-background" / "test_background_entities.tsv",
-                    "relations_file": None,
-                },
+                gen_kwargs={"data_dir": data_dir, "split": "development"},
             ),
         ]
 
-    def _generate_examples(self, **kwargs) -> Iterator[Tuple[str, Dict]]:
+    def _generate_examples(self, data_dir: Path, split: str) -> Iterator[Tuple[str, Dict]]:
         if self.config.name == "drugprot_source":
-            documents = self._read_source_examples(**kwargs)
+            documents = self._read_source_examples(data_dir, split)
             for document_id, document in documents.items():
                 yield document_id, document
 
         elif self.config.name == "drugprot_bigbio_kb":
-            documents = self._read_source_examples(**kwargs)
+            documents = self._read_source_examples(data_dir, split)
             for document_id, document in documents.items():
                 yield document_id, self._transform_source_to_kb(document)
 
-    def _read_source_examples(self, abstracts_file: Path, entities_file: Path, relations_file: Optional[Path]) -> Dict:
+    def _read_source_examples(self, input_dir: Path, split: str) -> Dict:
         """ """
-        # Note: The split "test-background" does not contain any relations
+        split_dir = input_dir / split
+        abstracts_file = split_dir / f"drugprot_{split}_abstracs.tsv"
+        entities_file = split_dir / f"drugprot_{split}_entities.tsv"
+        relations_file = split_dir / f"drugprot_{split}_relations.tsv"
+
         document_to_entities = collections.defaultdict(list)
         for line in entities_file.read_text().splitlines():
             columns = line.split("\t")
@@ -193,22 +181,20 @@ class DrugProtDataset(datasets.GeneratorBasedBuilder):
             )
 
         document_to_relations = collections.defaultdict(list)
+        for line in relations_file.read_text().splitlines():
+            columns = line.split("\t")
+            document_id = columns[0]
 
-        if relations_file is not None:
-            for line in relations_file.read_text().splitlines():
-                columns = line.split("\t")
-                document_id = columns[0]
+            document_relations = document_to_relations[document_id]
 
-                document_relations = document_to_relations[document_id]
-
-                document_relations.append(
-                    {
-                        "id": document_id + "_" + str(len(document_relations)),
-                        "type": columns[1],
-                        "arg1_id": document_id + "_" + columns[2][5:],
-                        "arg2_id": document_id + "_" + columns[3][5:],
-                    }
-                )
+            document_relations.append(
+                {
+                    "id": document_id + "_" + str(len(document_relations)),
+                    "type": columns[1],
+                    "arg1_id": document_id + "_" + columns[2][5:],
+                    "arg2_id": document_id + "_" + columns[3][5:],
+                }
+            )
 
         document_to_source = {}
         for line in abstracts_file.read_text().splitlines():
